@@ -2,7 +2,10 @@ import 'dart:async';
 import 'dart:developer';
 
 import 'package:easy_sidemenu/easy_sidemenu.dart';
-import 'package:flipper_dashboard/product_view.dart';
+import 'package:flipper_dashboard/functions.dart';
+import 'package:flipper_dashboard/layout.dart';
+import 'package:flipper_dashboard/pininput.dart';
+import 'package:flipper_dashboard/profile.dart';
 import 'package:flipper_models/isar_models.dart';
 import 'package:flipper_services/app_service.dart';
 import 'package:flipper_services/constants.dart';
@@ -12,12 +15,11 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_windowmanager/flutter_windowmanager.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:move_to_background/move_to_background.dart';
 import 'package:nfc_manager/nfc_manager.dart';
-import 'package:permission_handler/permission_handler.dart' as perm;
+import 'package:permission_handler/permission_handler.dart' as permission;
+import 'package:flutter/scheduler.dart';
+import 'package:pinput/pinput.dart';
 import 'package:stacked/stacked.dart';
-
-import 'page_switcher.dart';
 
 class FlipperApp extends StatefulWidget {
   const FlipperApp({Key? key}) : super(key: key);
@@ -31,6 +33,9 @@ class _FlipperAppState extends State<FlipperApp> with WidgetsBindingObserver {
   final TextEditingController controller = TextEditingController();
   SideMenuController sideMenu = SideMenuController();
   int tabselected = 0;
+  OverlayEntry? _overlayEntry;
+  final formKey = GlobalKey<FormState>();
+
   Future<void> _disableScreenshots() async {
     if (!kDebugMode && !isDesktopOrWeb) {
       await FlutterWindowManager.addFlags(FlutterWindowManager.FLAG_SECURE);
@@ -47,8 +52,8 @@ class _FlipperAppState extends State<FlipperApp> with WidgetsBindingObserver {
 
   @override
   void dispose() {
-    super.dispose();
     AppService.cleanedDataController.close();
+    super.dispose();
   }
 
   Future<void> nfc() async {
@@ -86,6 +91,88 @@ class _FlipperAppState extends State<FlipperApp> with WidgetsBindingObserver {
     }
   }
 
+  final defaultPinTheme = PinTheme(
+    width: 56,
+    height: 56,
+    textStyle: TextStyle(
+        fontSize: 20,
+        color: Color.fromRGBO(30, 60, 87, 1),
+        fontWeight: FontWeight.w600),
+    decoration: BoxDecoration(
+      border: Border.all(color: Color.fromRGBO(234, 239, 243, 1)),
+      borderRadius: BorderRadius.circular(20),
+    ),
+  );
+
+  void _insertOverlay(
+      {required BuildContext context, required HomeViewModel model}) {
+    _overlayEntry = OverlayEntry(builder: (context) {
+      final size = MediaQuery.of(context).size;
+      print(size.width);
+      return Material(
+        color: Colors.transparent,
+        child: Scaffold(
+          backgroundColor: Color(0xFF6F2F9).withOpacity(0.6),
+          resizeToAvoidBottomInset: true,
+          body: GestureDetector(
+            onTap: () {
+              // Handle tap on the overlay
+              print('ON TAP OVERLAY!');
+            },
+            child: Stack(
+              children: [
+                Positioned.fill(
+                  child: Align(
+                    alignment: Alignment.center,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        FutureBuilder<ITenant?>(
+                          future: ProxyService.isar.getTenantBYUserId(
+                            userId: ProxyService.box.getUserId()!,
+                          ),
+                          builder: (context, snapshot) {
+                            if (snapshot.connectionState ==
+                                    ConnectionState.waiting ||
+                                !snapshot.hasData) {
+                              return SizedBox.shrink();
+                            }
+                            ITenant tenant = snapshot.data!;
+                            return Padding(
+                              padding: const EdgeInsets.only(
+                                  bottom: 12.0), // Adjust spacing
+                              child: ProfileWidget(
+                                tenant: tenant,
+                                size: 25,
+                                sessionActive: tenant.sessionActive!,
+                                showIcon: false,
+                              ),
+                            );
+                          },
+                        ),
+                        Container(
+                          decoration: BoxDecoration(
+                              color: Color.fromARGB(255, 230, 230, 230),
+                              borderRadius: BorderRadius.circular(10)),
+                          width: 320,
+                          height: 140,
+                          child: OnlyBottomCursor(model: model),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      );
+    });
+
+    Overlay.of(context).insert(_overlayEntry!);
+  }
+
   @override
   Widget build(BuildContext context) {
     return ViewModelBuilder<HomeViewModel>.reactive(
@@ -94,6 +181,9 @@ class _FlipperAppState extends State<FlipperApp> with WidgetsBindingObserver {
         onViewModelReady: (model) async {
           //get default tenant
           model.defaultTenant();
+          ProxyService.isar.refreshSession(
+              branchId: ProxyService.box.getBranchId()!,
+              refreshRate: kDebugMode ? 1 : 5);
 
           /// if there is current order ongoing show them when the app starts
           ProxyService.dynamicLink.handleDynamicLink(context);
@@ -120,17 +210,16 @@ class _FlipperAppState extends State<FlipperApp> with WidgetsBindingObserver {
           model.loadReport();
           if (!isWindows) {
             await [
-              perm.Permission.storage,
-              perm.Permission.manageExternalStorage
+              permission.Permission.storage,
+              permission.Permission.manageExternalStorage
             ].request();
           }
         },
         builder: (context, model, child) {
           return WillPopScope(
             onWillPop: () async {
-              return _onWillPop(
-                model,
-              );
+              return onWillPop(
+                  context: context, message: 'Do you want to leave this app?');
             },
             child: Scaffold(
               appBar: AppBar(
@@ -150,159 +239,34 @@ class _FlipperAppState extends State<FlipperApp> with WidgetsBindingObserver {
                         ? 25
                         : 0,
               ),
-              body: LayoutBuilder(
-                  builder: (BuildContext context, BoxConstraints constraints) {
-                if (constraints.maxWidth < 600) {
-                  // this is a phone
-
-                  return PageSwitcher(
-                    controller: controller,
-                    model: model,
-                    currentPage: tabselected,
-                  );
-                } else {
-                  return Row(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: [
-                      SizedBox(
-                        width: 20,
-                        child: SizedBox.shrink(),
-                      ),
-                      // Left menu
-                      // ignore: todo
-                      // TODO: left menu will be essential when we add socials feature
-                      // by the time I implement it I will remove the above SizedBox as it is helping me
-                      // to keep the design consistent for now.
-                      // Container(
-                      //   width: 150,
-                      //   child: SideMenu(
-                      //     style: SideMenuStyle(
-                      //       showTooltip: false,
-                      //       displayMode: SideMenuDisplayMode.compact,
-                      //       compactSideMenuWidth: 60,
-                      //       openSideMenuWidth: 150,
-                      //     ),
-                      //     items: [
-                      //       SideMenuItem(
-                      //         // Priority of item to show on SideMenu, lower value is displayed at the top
-                      //         priority: 0,
-                      //         title: 'Dashboard',
-                      //         icon: Icon(Icons.home),
-                      //         badgeContent: Text(
-                      //           '3',
-                      //           style: TextStyle(color: Colors.white),
-                      //         ),
-                      //       )
-                      //     ],
-                      //     controller: sideMenu,
-                      //   ),
-                      // ),
-
-                      // Middle menu
-                      Expanded(
-                        flex: 2,
-                        child: Container(
-                          child: ProductView.normalMode(),
-                        ),
-                      ),
-
-                      // Right menu
-                      Expanded(
-                        flex: 1,
-                        child: Container(
-                          child: PageSwitcher(
-                            isBigScreen: true,
-                            controller: controller,
-                            model: model,
-                            currentPage: tabselected,
-                          ),
-                        ),
-                      ),
-                    ],
-                  );
-                }
-              }),
+              body: StreamBuilder<ITenant?>(
+                  stream: ProxyService.isar.authState(
+                    branchId: ProxyService.box.getBranchId()!,
+                  ),
+                  builder: (context, snapshot) {
+                    if (snapshot.hasData && !snapshot.data!.sessionActive!) {
+                      SchedulerBinding.instance.addPostFrameCallback((_) {
+                        _removeOverlay();
+                        _insertOverlay(context: context, model: model);
+                      });
+                    } else if (snapshot.hasData &&
+                        snapshot.data!.sessionActive!) {
+                      _removeOverlay();
+                    }
+                    return AppLayoutDraw(
+                        controller: controller,
+                        tabselected: tabselected,
+                        model: model);
+                  }),
             ),
           );
         });
   }
 
-  Future<bool> _onWillPop(HomeViewModel model) async {
-    final shouldPop = await showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text('Confirm'),
-          content: Text('Do you want to leave this app?'),
-          actions: <Widget>[
-            OutlinedButton(
-              child: Text('No',
-                  style: GoogleFonts.poppins(
-                    fontSize: 16.0,
-                    fontWeight: FontWeight.w500,
-                    color: Colors.white,
-                  )),
-              style: ButtonStyle(
-                backgroundColor:
-                    MaterialStateProperty.all<Color>(const Color(0xff006AFE)),
-                overlayColor: MaterialStateProperty.resolveWith<Color?>(
-                  (Set<MaterialState> states) {
-                    if (states.contains(MaterialState.hovered)) {
-                      return Colors.blue.withOpacity(0.04);
-                    }
-                    if (states.contains(MaterialState.focused) ||
-                        states.contains(MaterialState.pressed)) {
-                      return Colors.blue.withOpacity(0.12);
-                    }
-                    return null; // Defer to the widget's default.
-                  },
-                ),
-              ),
-              onPressed: () => Navigator.of(context).pop(false),
-            ),
-            OutlinedButton(
-              child: Text('Yes',
-                  style: GoogleFonts.poppins(
-                    fontSize: 16.0,
-                    fontWeight: FontWeight.w500,
-                    color: Colors.white,
-                  )),
-              style: ButtonStyle(
-                backgroundColor:
-                    MaterialStateProperty.all<Color>(const Color(0xff006AFE)),
-                overlayColor: MaterialStateProperty.resolveWith<Color?>(
-                  (Set<MaterialState> states) {
-                    if (states.contains(MaterialState.hovered)) {
-                      return Colors.blue.withOpacity(0.04);
-                    }
-                    if (states.contains(MaterialState.focused) ||
-                        states.contains(MaterialState.pressed)) {
-                      return Colors.blue.withOpacity(0.12);
-                    }
-                    return null; // Defer to the widget's default.
-                  },
-                ),
-              ),
-              onPressed: () => Navigator.of(context).pop(true),
-            ),
-          ],
-        );
-      },
-    );
-
-    if (shouldPop == true) {
-      /// @Decision: maybe a user does not want to close the day yet instead he want
-      /// to just send the app in background, hence why I have commented out
-      /// the code to trigger the closing drawer, for that
-      /// a user will have to be explicit
-      await MoveToBackground.moveTaskToBack();
-      // Drawers? drawer = await ProxyService.isar
-      //     .getDrawer(cashierId: ProxyService.box.getBusinessId()!);
-      // _routerService
-      //     .replaceWith(DrawerScreenRoute(open: "close", drawer: drawer));
-      return false;
-    } else {
-      return false;
+  void _removeOverlay() {
+    if (_overlayEntry != null) {
+      _overlayEntry?.remove();
+      _overlayEntry = null;
     }
   }
 }
