@@ -185,7 +185,7 @@ class _ScannViewState extends State<ScannView> {
         int businessId = ProxyService.box.getBusinessId()!;
         int branchId = ProxyService.box.getBranchId()!;
         String phone = ProxyService.box.getUserPhone()!;
-        int defaultApp = ProxyService.box.getDefaultApp();
+        String defaultApp = ProxyService.box.getDefaultApp();
         ProxyService.event.publish(loginDetails: {
           // this avoid users to log in to unintended device!
           'channel': split[1],
