@@ -1,7 +1,11 @@
+import 'dart:developer';
+
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 import 'package:firebase_ui_oauth_google/firebase_ui_oauth_google.dart';
 import 'package:flipper_routing/app.router.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 import 'package:flipper_routing/app.locator.dart';
@@ -59,8 +63,68 @@ class AuthOptionPage extends StatelessWidget {
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 20),
           child: AuthStateListener<OAuthController>(
-            child: OAuthProviderButton(
-              provider: GoogleProvider(clientId: _googleClientId),
+            child: Column(
+              children: [
+                SizedBox(
+                  width: 368,
+                  height: 68,
+                  child: OAuthProviderButton(
+                    variant: OAuthButtonVariant.icon,
+                    provider: GoogleProvider(clientId: _googleClientId),
+                  ),
+                ),
+                SizedBox(
+                    width: 368,
+                    height: 68,
+                    child: Container(
+                      decoration: BoxDecoration(
+                        boxShadow: [
+                          BoxShadow(
+                            color: Colors.grey.withOpacity(0.2),
+                            spreadRadius: 2,
+                            blurRadius: 5,
+                            offset: Offset(
+                                0, 3), // changes the position of the shadow
+                          ),
+                        ],
+                        borderRadius: BorderRadius.circular(
+                            4), // Add border radius to match IconButton's shape
+                      ),
+                      child: IconButton(
+                        onPressed: () async {
+                          log('microsoft');
+                          final provider = MicrosoftAuthProvider();
+                          provider.addScope('mail.read');
+                          final user =
+                              await FirebaseAuth.instance.signInWithProvider(
+                            provider,
+                          );
+                          if (user.user != null) {
+                            final _routerService = locator<RouterService>();
+                            _routerService.clearStackAndShow(
+                                StartUpViewRoute(invokeLogin: true));
+                          }
+                        },
+                        icon: SvgPicture.asset(
+                          'assets/microsoft.svg',
+                          package: 'flipper_login',
+                        ),
+                        style: primaryButtonStyle.copyWith(
+                            backgroundColor:
+                                MaterialStateProperty.all<Color>(Colors.white),
+                            side: MaterialStateProperty.resolveWith<BorderSide>(
+                                (states) => BorderSide(
+                                      color: Colors.grey.withOpacity(.1),
+                                    )),
+                            shape: MaterialStateProperty.resolveWith<
+                                    OutlinedBorder>(
+                                (states) => RoundedRectangleBorder(
+                                    borderRadius: BorderRadius.circular(10)))
+                            // You can remove the side property as the border is now managed by the Container's decoration
+                            ),
+                      ),
+                    ))
+              ],
             ),
             listener: (oldState, newState, ctrl) {
               if (newState is SignedIn) {
