@@ -9,20 +9,10 @@ part of 'business_type.dart';
 BusinessType _$BusinessTypeFromJson(Map<String, dynamic> json) => BusinessType(
       id: json['id'] as String,
       typeName: json['typeName'] as String,
-      deletedAt: json['deletedAt'] == null
-          ? null
-          : DateTime.parse(json['deletedAt'] as String),
-    )
-      ..action = json['action'] as String
-      ..lastTouched = json['lastTouched'] == null
-          ? null
-          : DateTime.parse(json['lastTouched'] as String);
+    );
 
 Map<String, dynamic> _$BusinessTypeToJson(BusinessType instance) =>
     <String, dynamic>{
-      'action': instance.action,
       'id': instance.id,
       'typeName': instance.typeName,
-      'lastTouched': instance.lastTouched?.toIso8601String(),
-      'deletedAt': instance.deletedAt?.toIso8601String(),
     };
