@@ -57,6 +57,10 @@ const ITenantSchema = IsarGeneratedSchema(
         type: IsarType.dateTime,
       ),
       IsarPropertySchema(
+        name: 'sessionActive',
+        type: IsarType.bool,
+      ),
+      IsarPropertySchema(
         name: 'action',
         type: IsarType.string,
       ),
@@ -101,7 +105,15 @@ int serializeITenant(IsarWriter writer, ITenant object) {
           -9223372036854775808);
   IsarCore.writeLong(writer, 9,
       object.deletedAt?.toUtc().microsecondsSinceEpoch ?? -9223372036854775808);
-  IsarCore.writeString(writer, 10, object.action);
+  {
+    final value = object.sessionActive;
+    if (value == null) {
+      IsarCore.writeNull(writer, 10);
+    } else {
+      IsarCore.writeBool(writer, 10, value);
+    }
+  }
+  IsarCore.writeString(writer, 11, object.action);
   return object.id;
 }
 
@@ -131,6 +143,14 @@ ITenant deserializeITenant(IsarReader reader) {
           DateTime.fromMicrosecondsSinceEpoch(value, isUtc: true).toLocal();
     }
   }
+  final bool? _sessionActive;
+  {
+    if (IsarCore.readNull(reader, 10)) {
+      _sessionActive = null;
+    } else {
+      _sessionActive = IsarCore.readBool(reader, 10);
+    }
+  }
   final object = ITenant(
     id: _id,
     name: _name,
@@ -140,6 +160,7 @@ ITenant deserializeITenant(IsarReader reader) {
     businessId: _businessId,
     userId: _userId,
     deletedAt: _deletedAt,
+    sessionActive: _sessionActive,
   );
   object.imageUrl = IsarCore.readString(reader, 7);
   {
@@ -151,7 +172,7 @@ ITenant deserializeITenant(IsarReader reader) {
           DateTime.fromMicrosecondsSinceEpoch(value, isUtc: true).toLocal();
     }
   }
-  object.action = IsarCore.readString(reader, 10) ?? '';
+  object.action = IsarCore.readString(reader, 11) ?? '';
   return object;
 }
 
@@ -195,7 +216,15 @@ dynamic deserializeITenantProp(IsarReader reader, int property) {
         }
       }
     case 10:
-      return IsarCore.readString(reader, 10) ?? '';
+      {
+        if (IsarCore.readNull(reader, 10)) {
+          return null;
+        } else {
+          return IsarCore.readBool(reader, 10);
+        }
+      }
+    case 11:
+      return IsarCore.readString(reader, 11) ?? '';
     default:
       throw ArgumentError('Unknown property: $property');
   }
@@ -213,6 +242,7 @@ sealed class _ITenantUpdate {
     String? imageUrl,
     DateTime? lastTouched,
     DateTime? deletedAt,
+    bool? sessionActive,
     String? action,
   });
 }
@@ -234,6 +264,7 @@ class _ITenantUpdateImpl implements _ITenantUpdate {
     Object? imageUrl = ignore,
     Object? lastTouched = ignore,
     Object? deletedAt = ignore,
+    Object? sessionActive = ignore,
     Object? action = ignore,
   }) {
     return collection.updateProperties([
@@ -248,7 +279,8 @@ class _ITenantUpdateImpl implements _ITenantUpdate {
           if (imageUrl != ignore) 7: imageUrl as String?,
           if (lastTouched != ignore) 8: lastTouched as DateTime?,
           if (deletedAt != ignore) 9: deletedAt as DateTime?,
-          if (action != ignore) 10: action as String?,
+          if (sessionActive != ignore) 10: sessionActive as bool?,
+          if (action != ignore) 11: action as String?,
         }) >
         0;
   }
@@ -266,6 +298,7 @@ sealed class _ITenantUpdateAll {
     String? imageUrl,
     DateTime? lastTouched,
     DateTime? deletedAt,
+    bool? sessionActive,
     String? action,
   });
 }
@@ -287,6 +320,7 @@ class _ITenantUpdateAllImpl implements _ITenantUpdateAll {
     Object? imageUrl = ignore,
     Object? lastTouched = ignore,
     Object? deletedAt = ignore,
+    Object? sessionActive = ignore,
     Object? action = ignore,
   }) {
     return collection.updateProperties(id, {
@@ -299,7 +333,8 @@ class _ITenantUpdateAllImpl implements _ITenantUpdateAll {
       if (imageUrl != ignore) 7: imageUrl as String?,
       if (lastTouched != ignore) 8: lastTouched as DateTime?,
       if (deletedAt != ignore) 9: deletedAt as DateTime?,
-      if (action != ignore) 10: action as String?,
+      if (sessionActive != ignore) 10: sessionActive as bool?,
+      if (action != ignore) 11: action as String?,
     });
   }
 }
@@ -321,6 +356,7 @@ sealed class _ITenantQueryUpdate {
     String? imageUrl,
     DateTime? lastTouched,
     DateTime? deletedAt,
+    bool? sessionActive,
     String? action,
   });
 }
@@ -342,6 +378,7 @@ class _ITenantQueryUpdateImpl implements _ITenantQueryUpdate {
     Object? imageUrl = ignore,
     Object? lastTouched = ignore,
     Object? deletedAt = ignore,
+    Object? sessionActive = ignore,
     Object? action = ignore,
   }) {
     return query.updateProperties(limit: limit, {
@@ -354,7 +391,8 @@ class _ITenantQueryUpdateImpl implements _ITenantQueryUpdate {
       if (imageUrl != ignore) 7: imageUrl as String?,
       if (lastTouched != ignore) 8: lastTouched as DateTime?,
       if (deletedAt != ignore) 9: deletedAt as DateTime?,
-      if (action != ignore) 10: action as String?,
+      if (sessionActive != ignore) 10: sessionActive as bool?,
+      if (action != ignore) 11: action as String?,
     });
   }
 }
@@ -383,6 +421,7 @@ class _ITenantQueryBuilderUpdateImpl implements _ITenantQueryUpdate {
     Object? imageUrl = ignore,
     Object? lastTouched = ignore,
     Object? deletedAt = ignore,
+    Object? sessionActive = ignore,
     Object? action = ignore,
   }) {
     final q = query.build();
@@ -397,7 +436,8 @@ class _ITenantQueryBuilderUpdateImpl implements _ITenantQueryUpdate {
         if (imageUrl != ignore) 7: imageUrl as String?,
         if (lastTouched != ignore) 8: lastTouched as DateTime?,
         if (deletedAt != ignore) 9: deletedAt as DateTime?,
-        if (action != ignore) 10: action as String?,
+        if (sessionActive != ignore) 10: sessionActive as bool?,
+        if (action != ignore) 11: action as String?,
       });
     } finally {
       q.close();
@@ -1578,6 +1618,32 @@ extension ITenantQueryFilter
     });
   }
 
+  QueryBuilder<ITenant, ITenant, QAfterFilterCondition> sessionActiveIsNull() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addFilterCondition(const IsNullCondition(property: 10));
+    });
+  }
+
+  QueryBuilder<ITenant, ITenant, QAfterFilterCondition>
+      sessionActiveIsNotNull() {
+    return QueryBuilder.apply(not(), (query) {
+      return query.addFilterCondition(const IsNullCondition(property: 10));
+    });
+  }
+
+  QueryBuilder<ITenant, ITenant, QAfterFilterCondition> sessionActiveEqualTo(
+    bool? value,
+  ) {
+    return QueryBuilder.apply(this, (query) {
+      return query.addFilterCondition(
+        EqualCondition(
+          property: 10,
+          value: value,
+        ),
+      );
+    });
+  }
+
   QueryBuilder<ITenant, ITenant, QAfterFilterCondition> actionEqualTo(
     String value, {
     bool caseSensitive = true,
@@ -1585,7 +1651,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         EqualCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1600,7 +1666,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         GreaterCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1616,7 +1682,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         GreaterOrEqualCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1631,7 +1697,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         LessCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1646,7 +1712,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         LessOrEqualCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1662,7 +1728,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         BetweenCondition(
-          property: 10,
+          property: 11,
           lower: lower,
           upper: upper,
           caseSensitive: caseSensitive,
@@ -1678,7 +1744,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         StartsWithCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1693,7 +1759,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         EndsWithCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1707,7 +1773,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         ContainsCondition(
-          property: 10,
+          property: 11,
           value: value,
           caseSensitive: caseSensitive,
         ),
@@ -1721,7 +1787,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         MatchesCondition(
-          property: 10,
+          property: 11,
           wildcard: pattern,
           caseSensitive: caseSensitive,
         ),
@@ -1733,7 +1799,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         const EqualCondition(
-          property: 10,
+          property: 11,
           value: '',
         ),
       );
@@ -1744,7 +1810,7 @@ extension ITenantQueryFilter
     return QueryBuilder.apply(this, (query) {
       return query.addFilterCondition(
         const GreaterCondition(
-          property: 10,
+          property: 11,
           value: '',
         ),
       );
@@ -1912,11 +1978,23 @@ extension ITenantQuerySortBy on QueryBuilder<ITenant, ITenant, QSortBy> {
     });
   }
 
+  QueryBuilder<ITenant, ITenant, QAfterSortBy> sortBySessionActive() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addSortBy(10);
+    });
+  }
+
+  QueryBuilder<ITenant, ITenant, QAfterSortBy> sortBySessionActiveDesc() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addSortBy(10, sort: Sort.desc);
+    });
+  }
+
   QueryBuilder<ITenant, ITenant, QAfterSortBy> sortByAction(
       {bool caseSensitive = true}) {
     return QueryBuilder.apply(this, (query) {
       return query.addSortBy(
-        10,
+        11,
         caseSensitive: caseSensitive,
       );
     });
@@ -1926,7 +2004,7 @@ extension ITenantQuerySortBy on QueryBuilder<ITenant, ITenant, QSortBy> {
       {bool caseSensitive = true}) {
     return QueryBuilder.apply(this, (query) {
       return query.addSortBy(
-        10,
+        11,
         sort: Sort.desc,
         caseSensitive: caseSensitive,
       );
@@ -2064,17 +2142,29 @@ extension ITenantQuerySortThenBy
     });
   }
 
+  QueryBuilder<ITenant, ITenant, QAfterSortBy> thenBySessionActive() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addSortBy(10);
+    });
+  }
+
+  QueryBuilder<ITenant, ITenant, QAfterSortBy> thenBySessionActiveDesc() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addSortBy(10, sort: Sort.desc);
+    });
+  }
+
   QueryBuilder<ITenant, ITenant, QAfterSortBy> thenByAction(
       {bool caseSensitive = true}) {
     return QueryBuilder.apply(this, (query) {
-      return query.addSortBy(10, caseSensitive: caseSensitive);
+      return query.addSortBy(11, caseSensitive: caseSensitive);
     });
   }
 
   QueryBuilder<ITenant, ITenant, QAfterSortBy> thenByActionDesc(
       {bool caseSensitive = true}) {
     return QueryBuilder.apply(this, (query) {
-      return query.addSortBy(10, sort: Sort.desc, caseSensitive: caseSensitive);
+      return query.addSortBy(11, sort: Sort.desc, caseSensitive: caseSensitive);
     });
   }
 }
@@ -2139,10 +2229,16 @@ extension ITenantQueryWhereDistinct
     });
   }
 
+  QueryBuilder<ITenant, ITenant, QAfterDistinct> distinctBySessionActive() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addDistinctBy(10);
+    });
+  }
+
   QueryBuilder<ITenant, ITenant, QAfterDistinct> distinctByAction(
       {bool caseSensitive = true}) {
     return QueryBuilder.apply(this, (query) {
-      return query.addDistinctBy(10, caseSensitive: caseSensitive);
+      return query.addDistinctBy(11, caseSensitive: caseSensitive);
     });
   }
 }
@@ -2208,9 +2304,15 @@ extension ITenantQueryProperty1 on QueryBuilder<ITenant, ITenant, QProperty> {
     });
   }
 
-  QueryBuilder<ITenant, String, QAfterProperty> actionProperty() {
+  QueryBuilder<ITenant, bool?, QAfterProperty> sessionActiveProperty() {
     return QueryBuilder.apply(this, (query) {
       return query.addProperty(10);
+    });
+  }
+
+  QueryBuilder<ITenant, String, QAfterProperty> actionProperty() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addProperty(11);
     });
   }
 }
@@ -2276,9 +2378,15 @@ extension ITenantQueryProperty2<R> on QueryBuilder<ITenant, R, QAfterProperty> {
     });
   }
 
-  QueryBuilder<ITenant, (R, String), QAfterProperty> actionProperty() {
+  QueryBuilder<ITenant, (R, bool?), QAfterProperty> sessionActiveProperty() {
     return QueryBuilder.apply(this, (query) {
       return query.addProperty(10);
+    });
+  }
+
+  QueryBuilder<ITenant, (R, String), QAfterProperty> actionProperty() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addProperty(11);
     });
   }
 }
@@ -2346,9 +2454,15 @@ extension ITenantQueryProperty3<R1, R2>
     });
   }
 
-  QueryBuilder<ITenant, (R1, R2, String), QOperations> actionProperty() {
+  QueryBuilder<ITenant, (R1, R2, bool?), QOperations> sessionActiveProperty() {
     return QueryBuilder.apply(this, (query) {
       return query.addProperty(10);
+    });
+  }
+
+  QueryBuilder<ITenant, (R1, R2, String), QOperations> actionProperty() {
+    return QueryBuilder.apply(this, (query) {
+      return query.addProperty(11);
     });
   }
 }
@@ -2368,6 +2482,7 @@ ITenant _$ITenantFromJson(Map<String, dynamic> json) => ITenant(
       deletedAt: json['deletedAt'] == null
           ? null
           : DateTime.parse(json['deletedAt'] as String),
+      sessionActive: json['sessionActive'] as bool?,
     )
       ..action = json['action'] as String
       ..imageUrl = json['imageUrl'] as String?
@@ -2387,4 +2502,5 @@ Map<String, dynamic> _$ITenantToJson(ITenant instance) => <String, dynamic>{
       'imageUrl': instance.imageUrl,
       'lastTouched': instance.lastTouched?.toIso8601String(),
       'deletedAt': instance.deletedAt?.toIso8601String(),
+      'sessionActive': instance.sessionActive,
     };
