@@ -72,10 +72,14 @@ class IsarAPI<M> implements IsarApiInterface {
           ConversationSchema,
           DeviceSchema,
           FavoriteSchema,
-          EBMSchema
+          EBMSchema,
+          UserActivitySchema
         ],
         directory: foundation.kIsWeb ? Isar.sqliteInMemory : appDocDir.path,
-        engine: foundation.kIsWeb ? IsarEngine.sqlite : IsarEngine.isar,
+        engine: foundation.kIsWeb || Platform.isLinux
+            ? IsarEngine.sqlite
+            : IsarEngine.isar,
+        name: 'flipper-db',
       );
     } else {
       db = isa;
@@ -146,7 +150,7 @@ class IsarAPI<M> implements IsarApiInterface {
         isar.transactions.put(transaction);
       });
       Transaction? createdTransaction = await db.transactions.get(id);
-      ProxyService.box.write(key: 'currentTransactionId', value: id);
+      ProxyService.box.writeString(key: 'currentTransactionId', value: id);
       return createdTransaction!;
     } else {
       return existTransaction;
@@ -185,7 +189,7 @@ class IsarAPI<M> implements IsarApiInterface {
         isar.transactions.put(transaction);
       });
       Transaction? createdTransaction = db.transactions.get(id);
-      ProxyService.box.write(key: 'currentTransactionId', value: id);
+      ProxyService.box.writeString(key: 'currentTransactionId', value: id);
       return createdTransaction!;
     } else {
       return existTransaction;
@@ -673,7 +677,7 @@ class IsarAPI<M> implements IsarApiInterface {
     int branchId = ProxyService.box.getBranchId()!;
     int businessId = ProxyService.box.getBusinessId()!;
     String phoneNumber = ProxyService.box.getUserPhone()!;
-    int defaultApp = ProxyService.box.getDefaultApp();
+    String defaultApp = ProxyService.box.getDefaultApp();
     final http.Response response = await flipperHttpClient.post(
       Uri.parse("$apihub/v2/api/pin"),
       body: jsonEncode(
@@ -1215,7 +1219,7 @@ class IsarAPI<M> implements IsarApiInterface {
         return isar.business.where().userIdEqualTo(userId).findFirst();
       });
     }
-    ProxyService.box.write(key: 'businessId', value: business!.id);
+    ProxyService.box.writeInt(key: 'businessId', value: business!.id);
 
     return business;
   }
@@ -1466,61 +1470,62 @@ class IsarAPI<M> implements IsarApiInterface {
     );
 
     if (response.statusCode == 200 && response.body.isNotEmpty) {
-      IUser syncF = IUser.fromRawJson(response.body);
-      await ProxyService.box.write(
+      IUser user = IUser.fromRawJson(response.body);
+      await ProxyService.box.writeString(
         key: 'userPhone',
         value: userPhone,
       );
-      await ProxyService.box.write(
+      await ProxyService.box.writeString(
         key: 'bearerToken',
-        value: syncF.token,
+        value: user.token,
       );
-      await ProxyService.box.write(
+      await ProxyService.box.writeInt(
         key: 'userId',
-        value: syncF.id,
+        value: user.id,
       );
-      await ProxyService.box.write(
+      recordUserActivity(userId: user.id, activity: 'login');
+      if (user.tenants.isEmpty) {
+        throw BusinessNotFoundException(
+            term:
+                "No tenant added to the user, if a business is added it should have one tenant");
+      }
+      await ProxyService.box.writeInt(
         key: 'branchId',
         // check if branches is empty
-        value: syncF.tenants.isEmpty
-            ? null
-            : syncF.tenants.first.branches.first.id,
+        value: user.tenants.isEmpty ? 0 : user.tenants.first.branches.first.id,
       );
-      await ProxyService.box.write(
+
+      log(user.id.toString(), name: 'login');
+      await ProxyService.box.writeInt(
         key: 'businessId',
         // check if businesses is empty
-        value: syncF.tenants.isEmpty
-            ? null
-            : syncF.tenants.first.businesses.first.id,
+        value:
+            user.tenants.isEmpty ? 0 : user.tenants.first.businesses.first.id,
       );
       if (skipDefaultAppSetup == false) {
-        await ProxyService.box.write(
+        await ProxyService.box.writeString(
           key: 'defaultApp',
 
           /// because we don update default app from server
           /// because we want the ability of switching apps to be entirely offline
           /// then if we have a default app in the box we use it if it only different from 1
-          value: syncF.tenants.isEmpty
-              ? null
-              : ProxyService.box.getDefaultApp() != 1
+          value: user.tenants.isEmpty
+              ? 'null'
+              : ProxyService.box.getDefaultApp() != "1"
                   ? ProxyService.box.getDefaultApp()
-                  : syncF.tenants.first.businesses.first.businessTypeId,
+                  : user.tenants.first.businesses.first.businessTypeId
+                      .toString(),
         );
       }
 
-      if (syncF.tenants.isEmpty) {
-        throw BusinessNotFoundException(
-            term:
-                "No tenant added to the user, if a business is added it should have one tenant");
-      }
-      for (Tenant tenant in syncF.tenants) {
+      for (Tenant tenant in user.tenants) {
         ITenant iTenant = ITenant(
             id: tenant.id,
             name: tenant.name,
             businessId: tenant.businessId,
             nfcEnabled: tenant.nfcEnabled,
             email: tenant.email,
-            userId: syncF.id,
+            userId: user.id,
             phoneNumber: tenant.phoneNumber);
 
         db.write((isar) {
@@ -1532,11 +1537,17 @@ class IsarAPI<M> implements IsarApiInterface {
         db.write((isar) {
           isar.permissions.putAll(tenant.permissions);
         });
+
         db.write((isar) {
-          isar.iTenants.put(iTenant);
+          if (user.id == iTenant.userId) {
+            iTenant.sessionActive = true;
+            isar.iTenants.put(iTenant);
+          } else {
+            isar.iTenants.put(iTenant);
+          }
         });
       }
-      return syncF;
+      return user;
     } else if (response.statusCode == 401) {
       throw SessionException(term: "session expired");
     } else if (response.statusCode == 500) {
@@ -1673,6 +1684,11 @@ class IsarAPI<M> implements IsarApiInterface {
 
   @override
   Future<T?> create<T>({required T data}) async {
+    /// update user activity model
+    int userId = ProxyService.box.getUserId()!;
+    recordUserActivity(userId: userId, activity: 'create');
+
+    /// end with updating user activity
     if (data is Conversation) {
       Conversation conversation = data;
       db.write((isar) {
@@ -1782,6 +1798,11 @@ class IsarAPI<M> implements IsarApiInterface {
   /// @Deprecated [endpoint] don't give the endpoint params
   @override
   Future<T?> update<T>({required T data}) async {
+    /// update user activity
+    int userId = ProxyService.box.getUserId()!;
+    recordUserActivity(userId: userId, activity: 'create');
+
+    /// end updating user activity
     // int branchId = ProxyService.box.getBranchId()!;
     if (data is Device) {
       Device device = data;
@@ -1854,7 +1875,8 @@ class IsarAPI<M> implements IsarApiInterface {
     if (data is EBM) {
       final ebm = data;
       db.write((isar) {
-        ProxyService.box.write(key: "serverUrl", value: ebm.taxServerUrl);
+        ProxyService.box
+            .writeString(key: "serverUrl", value: ebm.taxServerUrl ?? 'null');
         Business? business =
             isar.business.where().userIdEqualTo(ebm.userId).findFirst();
         business
@@ -2305,10 +2327,8 @@ class IsarAPI<M> implements IsarApiInterface {
   @override
   Future<List<BusinessType>> businessTypes() async {
     final responseJson = [
-      {"id": 1, "typeName": "Flipper Retailer"},
-      {"id": 2, "typeName": "Flipper Connecta"},
-      // {"id": 3, "typeName": "Retailer"},
-      // {"id": 4, "typeName": "Agent"}
+      {"id": "1", "typeName": "Flipper Retailer"},
+      {"id": "2", "typeName": "Flipper Connecta"},
     ];
     Future.delayed(Duration(seconds: 5));
     final response = http.Response(jsonEncode(responseJson), 200);
@@ -2426,8 +2446,12 @@ class IsarAPI<M> implements IsarApiInterface {
 
     final now = DateTime.now();
 
-    if (now.isBefore(token.validFrom) || now.isAfter(token.validUntil)) {
-      db.write((isar) => isar.tokens.delete(token.id));
+    if (token.validFrom != null && token.validUntil != null) {
+      if (now.isBefore(token.validFrom!) || now.isAfter(token.validUntil!)) {
+        db.write((isar) => isar.tokens.delete(token.id));
+        return false;
+      }
+    } else {
       return false;
     }
     String? localToken = ProxyService.box.whatsAppToken();
@@ -2438,7 +2462,7 @@ class IsarAPI<M> implements IsarApiInterface {
   }
 
   @override
-  Future<SocialToken> loginOnSocial(
+  Future<SocialToken?> loginOnSocial(
       {String? phoneNumberOrEmail, String? password}) async {
     final http.Response response = await socialsHttpClient.post(
       Uri.parse("$commApi/login"),
@@ -2449,7 +2473,10 @@ class IsarAPI<M> implements IsarApiInterface {
       SocialToken responseBody = SocialToken.fromRawJson(response.body);
       return responseBody;
     } else {
-      throw Exception("Failed to get token");
+      //// TODO: this is not a stopper when performing app global login
+      /// instead implement global non fatal error reporting so it can be addressed separately
+      // throw Exception("Failed to get token");
+      return null;
     }
   }
 
@@ -2611,12 +2638,12 @@ class IsarAPI<M> implements IsarApiInterface {
       }
 
       if (jsonDecode(response.body)['lastEvaluatedKey'] != null) {
-        // Set lastKey to the value returned by the API
+        /// Set lastKey to the value returned by the API
         String createdAt = jsonDecode(response.body)['lastEvaluatedKey']
             ['createdAt'] as String;
         int id = jsonDecode(response.body)['lastEvaluatedKey']['id'] as int;
-        ProxyService.box.write(key: 'createdAt', value: createdAt);
-        ProxyService.box.write(key: 'id', value: id);
+        ProxyService.box.writeString(key: 'createdAt', value: createdAt);
+        ProxyService.box.writeInt(key: 'id', value: id);
       } else {
         log(jsonDecode(response.body).toString());
         log("there is no last key to use in query and that is fine!");
@@ -3075,6 +3102,105 @@ class IsarAPI<M> implements IsarApiInterface {
           .watch(fireImmediately: true)
           .asyncMap((event) => event.first));
     }
+  }
+
+  Future<bool> hasNoActivityInLast5Minutes(
+      {required int userId, int? refreshRate = 5}) async {
+    // Get the current time
+    DateTime currentTime = DateTime.now();
+
+    // Calculate the time [timer] minutes ago
+    DateTime fiveMinutesAgo =
+        currentTime.subtract(Duration(minutes: refreshRate!));
+
+    // Retrieve the user activities
+    List<UserActivity> userActivities = await activities(userId: userId);
+
+    // Check if any activity was touched within the last 5 minutes
+    for (var activity in userActivities) {
+      if (activity.lastTouched!.isAfter(fiveMinutesAgo)) {
+        // The user has done an activity within the last 5 minutes
+        return false;
+      }
+    }
+
+    // No activity found within the last 5 minutes
+    return true;
+  }
+
+  @override
+  Stream<ITenant?> authState({required int branchId}) {
+    int userId = ProxyService.box.getUserId()!;
+    return db.read((isar) => isar.iTenants
+        .where()
+        .userIdEqualTo(userId)
+        .deletedAtIsNull()
+        .sortByLastTouchedDesc()
+        .watch(fireImmediately: true)
+        .asyncMap((event) => event.first));
+  }
+
+  @override
+  Future<void> refreshSession(
+      {required int branchId, int? refreshRate = 5}) async {
+    while (true) {
+      try {
+        int userId = ProxyService.box.getUserId()!;
+        bool session = await hasNoActivityInLast5Minutes(
+            userId: userId, refreshRate: refreshRate);
+        log(session.toString(), name: 'session');
+        log(userId.toString(), name: 'session');
+        if (!session) {
+          ITenant? tenant = await ProxyService.isar
+              .getTenantBYUserId(userId: ProxyService.box.getUserId()!);
+          tenant?.sessionActive = session;
+          ProxyService.isar.update(data: tenant);
+        }
+      } catch (error) {
+        print('Error fetching tenant: $error');
+      }
+      await Future.delayed(Duration(minutes: refreshRate!));
+    }
+  }
+
+  @override
+  Future<List<UserActivity>> activities({required int userId}) async {
+    // Get the current date
+    DateTime now = DateTime.now();
+
+    // Calculate the start and end of the current day
+    DateTime startOfDay = DateTime(now.year, now.month, now.day);
+    DateTime endOfDay = startOfDay.add(Duration(days: 1));
+
+    return db.read((isar) => isar.userActivitys
+        .where()
+        .lastTouchedBetween(startOfDay, endOfDay)
+        .findAll());
+  }
+
+  @override
+  Future<void> recordUserActivity(
+      {required int userId, required String activity}) async {
+    db.write((isar) => isar.userActivitys.put(UserActivity(
+        userId: userId,
+        action: activity,
+        id: randomString(),
+        timestamp: DateTime.now(),
+        lastTouched: DateTime.now())));
+  }
+
+  /// there is cases where singup fail and there is some variable saved
+  /// in this case for us to be clean better clean them for next attempt
+  @override
+  Future<void> logOutLight() async {
+    ProxyService.box.remove(key: 'userId');
+    ProxyService.box.remove(key: 'getIsTokenRegistered');
+    ProxyService.box.remove(key: 'bearerToken');
+    ProxyService.box.remove(key: 'branchId');
+    ProxyService.box.remove(key: 'userPhone');
+    ProxyService.box.remove(key: 'UToken');
+    ProxyService.box.remove(key: 'businessId');
+    ProxyService.box.remove(key: 'defaultApp');
   }
 
   /// End of streams
