@@ -149,8 +149,13 @@ class App {}
 // https://codepush.dev/
 //TODO: adapt to new FIGMA things https://uxplanet.org/whats-new-in-figma-10-updates-from-config-2023-c1651012835
 // https://codelabs.developers.google.com/design-android-launcher#4
-
+/// there was a time I used this frag in CMAKELists.txt and this was when
+///TODO: firebase core was not friendy and for now it seem to be fixed.
+// set(CMAKE_BUILD_TYPE "Release")
 // TODO: remove these deprecated fields from remote db
+// https://pub.dev/packages/pinput
 // active
 // reported
 // draft
+// getting key hash key
+//  keytool -list -v -keystore ~/.android/debug.keystore -alias androiddebugkey -storepass android -keypass android
