@@ -1,7 +1,11 @@
 abstract class LocalStorage {
-  dynamic read({required String key});
+  int? readInt({required String key});
+  String? readString({required String key});
+  bool? readBool({required String key});
   dynamic remove({required String key});
-  Future<bool> write({required String key, required dynamic value});
+  Future<bool> writeInt({required String key, required int value});
+  Future<bool> writeString({required String key, required String value});
+  Future<bool> writeBool({required String key, required bool value});
   int? getBusinessId();
   int? getBranchId();
   bool? getIsTokenRegistered();
@@ -18,7 +22,7 @@ abstract class LocalStorage {
   bool hasSignedInForAutoBackup();
   String? gdID();
   String? getBearerToken();
-  int getDefaultApp();
+  String getDefaultApp();
   String? whatsAppToken();
   String? paginationCreatedAt();
   int? paginationId();
