@@ -84,9 +84,9 @@ class FirebaseMessagingService implements Messaging {
           },
           collectionName: 'messagings',
         );
-        ProxyService.box.write(key: 'getIsTokenRegistered', value: true);
+        ProxyService.box.writeBool(key: 'getIsTokenRegistered', value: true);
       } catch (e) {
-        ProxyService.box.write(key: 'getIsTokenRegistered', value: true);
+        ProxyService.box.writeBool(key: 'getIsTokenRegistered', value: true);
       }
     }
   }
